+++ plugins/cost-insights/src/components/ProductInsightsCard/ProductInsightsCard.test.tsx
@@ -109,7 +109,7 @@ describe('<ProductInsightsCard/>', () => {
       aggregation: [0, 0],
       change: { ratio: 0, amount: 0 },
     };
-    const subheader = `There are no ${MockComputeEngine.name} costs within this timeframe for your team's projects.`;
+    const subheader = `There are no ${MockComputeEngine.name} costs within this time frame for your team's projects.`;
     const rendered = await renderProductInsightsCardInTestApp(
       entity,
       MockComputeEngine,