+++ plugins/techdocs-backend/src/service/standaloneServer.ts
@@ -52,14 +52,14 @@ export async function startStandaloneServer(
 
   logger.debug('Creating application...');
   const preparers = new Preparers();
-  const directoryPreparer = new DirectoryPreparer(logger);
+  const directoryPreparer = new DirectoryPreparer(config, logger);
   preparers.register('dir', directoryPreparer);
 
   const generators = new Generators();
   const techdocsGenerator = new TechdocsGenerator(logger, config);
   generators.register('techdocs', techdocsGenerator);
 
-  const publisher = Publisher.fromConfig(config, logger, discovery);
+  const publisher = await Publisher.fromConfig(config, { logger, discovery });
 
   const dockerClient = new Docker();