+++ packages/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/core
 
+## 0.4.3
+
+### Patch Changes
+
+- a08c32ced: Add `FlatRoutes` component to replace the top-level `Routes` component from `react-router` within apps, removing the need for manually appending `/*` to paths or sorting routes.
+- Updated dependencies [a08c32ced]
+- Updated dependencies [86c3c652a]
+- Updated dependencies [27f2af935]
+  - @backstage/core-api@0.2.8
+
+## 0.4.2
+
+### Patch Changes
+
+- 1dc445e89: Update to use new plugin extension API
+- 342270e4d: Create AboutCard in core and use it in pagerduty and catalog plugin
+- Updated dependencies [d681db2b5]
+- Updated dependencies [1dc445e89]
+  - @backstage/core-api@0.2.7
+
 ## 0.4.1
 
 ### Patch Changes