+++ plugins/graphiql/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-graphiql
 
+## 0.2.4
+
+### Patch Changes
+
+- 21f6be0f6: Patch release to force a release for the `esm.js` extension change for dynamically loading chunks
+
+## 0.2.3
+
+### Patch Changes
+
+- 1dc445e89: Update to use new plugin extension API
+- Updated dependencies [1dc445e89]
+- Updated dependencies [342270e4d]
+  - @backstage/core@0.4.2
+
 ## 0.2.2
 
 ### Patch Changes