+++ plugins/scaffolder-backend/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @backstage/plugin-scaffolder-backend
 
+## 0.4.0
+
+### Minor Changes
+
+- 5eb8c9b9e: Fix gitlab scaffolder publisher
+
+### Patch Changes
+
+- 7e3451700: bug(scaffolder): Ignore the .git folder when adding dot-files to the index
+
+## 0.3.7
+
+### Patch Changes
+
+- 37a5244ef: Add scaffolding support for Bitbucket Cloud and Server.
+- 00042e73c: Moving the Git actions to isomorphic-git instead of the node binding version of nodegit
+- 9efbc5585: Add config schema for Bitbucket scaffolder
+- Updated dependencies [5ecd50f8a]
+- Updated dependencies [00042e73c]
+- Updated dependencies [0829ff126]
+- Updated dependencies [036a84373]
+  - @backstage/backend-common@0.4.2
+  - @backstage/integration@0.1.5
+
+## 0.3.6
+
+### Patch Changes
+
+- 19554f6d6: Added Github Actions for Create React App, and allow better imports of files inside a module when they're exposed using `files` in `package.json`
+- 33a82a713: GitLab preparer uses the right token (primarily the same one as the publisher, falling back to the integrations token)
+- aed8f7f12: Clearer error message when preparer or publisher type can't be determined.
+
 ## 0.3.5
 
 ### Patch Changes